+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes to this project are documented in this file. More detail (including info
 
 ### Fixed
 - AUTOTOOL-3915: sourceAddress property ignored on non-internal virtuals
+- AUTOTOOL-3948: Not able to remove pool member's metadata from declaration
 
 ### Changed